+++ src/commands/taskCommands.ts
@@ -21,7 +21,10 @@ const fieldMap = new Map([
     ["Command", 255],
     ["STDOUT", 255],
     ["STDERR", 255],
-    ["Reason", 50]
+    ["Reason", 50],
+    ["SubmitTime", 20],
+    ["StartTime", 20],
+    ["EndTime", 20],
 ]);
 
 let cachePath: string;
@@ -40,7 +43,7 @@ function extractTask(taskString: string): Task[] {
     const decoder = new TextDecoder('utf-8');
     let taskList: Task[] = [];
     taskString.split('\n').forEach((value) => {
-        // JobID:20,Name:100,Username:50,State:20,NodeList:50,Gres:50,TimeLimit:20,TimeUsed:20,Command:255,STDOUT:255,STDERR:255,Reason:50
+        // JobID:20,Name:100,Username:50,State:20,NodeList:50,Gres:50,TimeLimit:20,TimeUsed:20,Command:255,STDOUT:255,STDERR:255,Reason:50,SubmitTime:20,StartTime:20,EndTime:20
         if (value.length === 0) { return; }
 
         const encodedValue = encoder.encode(value);
@@ -48,7 +51,7 @@ function extractTask(taskString: string): Task[] {
             arr.push(decoder.decode(encodedValue.slice(slices[i], v)).trim());
             return arr;
         }, []);
-        const task = new Task(fields[0], fields[1], fields[2], fields[3], fields[4], fields[5], fields[6], fields[7], fields[8], fields[9], fields[10], fields[11]);
+        const task = new Task(fields[0], fields[1], fields[2], fields[3], fields[4], fields[5], fields[6], fields[7], fields[8], fields[9], fields[10], fields[11], fields[12], fields[13], fields[14]);
         taskList.push(task);
     });
     return taskList;
@@ -76,7 +79,7 @@ async function refreshUserTasks() {
         vscode.window.showErrorMessage(err);
         return;
     }
-    taskService.updateTask(...extractTask(out));
+    await taskService.updateTask(...extractTask(out));
     taskView.taskViewDataProvider.refresh();
 }